Working on fatfs compiled and added control ctrl

+++ src/env_68k/src/stdlib/strntoumax.c
@@ -12,7 +12,8 @@ static inline int digitval(int ch)
     if (d < 6) return (int)(d+10);
     return -1;
 }
-uintmax_t strntoumax(const unsigned char *nptr, char **endptr, int base, size_t n){
+
+uintmax_t strntoumax(const unsigned char *nptr, char **endptr, int base, unsigned int n){
     const unsigned char*  p   = nptr;
     const unsigned char*  end = p + n;
     int                   minus = 0;
